+++ pkg/project/project.go
@@ -1,7 +1,7 @@
 package project
 
 var (
-	version = "2.6.2-dev"
+	version = "2.6.2"
 )
 
 func Version() string {